New script: cidrblocks
cidrblocks will output a list of all CIDR blocks that contain the
supplied IP address. Begin with /32 and return incrementally bigger
CIDR blocks until the stop limit is reach (a default is set in the
script, or it can be specified on the command line.

Examples:

cidrblocks 192.168.8.8		use the default CIDR mask
cidrblocks 192.168.8.8 24	stop at /24 instead of the default