switch cidrblocks back to sipcalc to catch/32
I was trying ipcalc as an alternative to sipcalc to find the network address.
Turns out that ipcalc doesn't think that /32 is a network (and it's right), but
we need to see /32 netmasks for our search to work. So, I'm switching back
to sipcalc, and moving to an awk-based capture of the ip address. Works fine.