ports-by-ip: consider junos defaults when greping
When trying to resolve a set of applications into a list of ports, we
need to look at applications that are defined by default in the
junos-defaults group. This requires us to fetch junos-defaults into a
separate file and include it in our greps.

Beause set groups and apply groups are built with wildcards and have
extra keywords in their set commands, we need to alter the grep regex to
be a little looser to accomodate the two different files.