Handle "any" addresses, only show "permit" policies
Some policies use (source|destination)-address "any" instead of an
address book entry. We need to catch those.

In many cases, "any" policies are actually explicit deny policies
instead of permits. So we need to look for "then permit" in the
appropriate place and return only policies that permit traffic.