cidrblocks cleanup, docs, standard CIDR notation
The cidrblocks script was one of the earliest things I wrote in this
suite, and since it's simple, I hadn't revisited it to really make it a
serious piece of code.

Add usage, better comments. Also, switch to using the standard IP/MASK
notation instead of "IP MASK" as I had it before. Also, added error
checking to make sure the request is at least close to a valid IP/MASK
combination.

Using sipcalc only now. Removed references to the option to use ipcalc,
as ipcalc doesn't display /32 at all.

Also in this update, a small change to policies-by-ip so it calls
cidrblocks with IP/MASK.