policies-by-ip: list "any" policies in correct order
Previously, policies with destination-address any came at the end of the
polcy list, rather than in the order of their evaluation, like the rest
of the polcies. This commit alters the destination-address grep so that
the any policies are found along with the policies reflecting a specific
IP address.

This is better, but probably still not right. We need to do a search for
all address book patterns at once, instead of iterating over them the
way we do now.

Now, things are searched like this: (pattern|all), then (pattern2|all),
etc. We need to search for (pattern1|pattern2|...|all) instead. This
will speed up the script and make the search more accurate.