policies-by-ip: remove all grep loops
OK, the loop has been optimised to a simple for .. echo .. done loop
now. The result is that there are only 3 greps in the actual search now,
instead of the earlier 3N greps (where N=number of CIDRs searched).
The three greps, do, in order:

1. Find the address-book objects for each CIDR
2. Find the policy names based on those objects.
3. Confirm that the policies are permit policies. *
(* we're still only interested in policies that open ports.)

This update makes the script run much faster, and it should be usable
for everyone now.

There are two other greps earlier on, at the error check stage, but as
they stop after the first search hit, they're much less intensive.