policies-by-ip: match any even when IP not matched specifically
Turns out that IP addresses had to show up somewhere in the address
book in order to be found. If an address (like a /32) wasn't listed
at all, the any match wouldn't be applied.

The change adapts the awk to better match the pattern and include
"any," and also makes the match a little faster.